2024/python_training/batch/pyproject.toml
[tool.poetry]
name = "person-similarity-batch"
version = "0.1.0"
description = "search person similarity batch"
authors = ["Sansan, Inc."]
license = "Proprietary"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
gokart = "^1.2.7"
pecanpy = "2.0.1"
scikit-learn = "^1.4.1.post1"
s3path = "^0.5.2"
s3fs = "^2024.2.0"
typing-extensions = "^4.10.0"
python-dotenv = "^1.0.1"
pandas-gbq = "^0.22.0"
gensim = "^4.3.2"
scipy = "1.12.0"


[tool.poetry.group.dev.dependencies]
pandas-stubs = "^2.2.1.240316"
mypy = "^1.5.1"
pytest = "^7.4.2"
ruff = "^0.3.4"
faker = "^19.6.2"
mypy-extensions = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
check_untyped_defs = true
ignore_errors = false
strict_optional = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 120
select = ["ALL"]
ignore = [
    # missing-trailing-comma
    "COM812",
    # ambiguous-variable-name
    "E741",
    #  single-line-implicit-string-concatenation
    "ISC001",
    # assert
    "S101",
    # suspicious-non-cryptographic-random-usage
    "S311",
    # unnecessary-assign
    "RET504",
    # magic-value-comparison
    "PLR2004",
    # pydocstyle
    "D",
    # flake8-annotations
    "ANN",
]
exclude = [
    ".git",
    ".github",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
]

[tool.ruff.isort]
known-first-party = ["pipeline"]

[tool.ruff.mccabe]
max-complexity = 5
